@@ -1,63 +1,59 @@
 {
-    "name": "friendsofsymfony/oauth-server-bundle",
-    "type": "symfony-bundle",
-    "description": "Symfony2 OAuth Server Bundle",
-    "keywords": ["oauth", "oauth2", "server"],
-    "homepage": "http://friendsofsymfony.github.com",
-    "license": "MIT",
-    "authors": [
-        {
-          "name": "Arnaud Le Blanc",
-          "email": "arnaud.lb@gmail.com"
-        },
-        {
-            "name": "FriendsOfSymfony Community",
-            "homepage": "https://github.com/FriendsOfSymfony/FOSOAuthServerBundle/contributors"
-        }
-    ],
-    "require": {
-        "php": "^5.5|^7.0",
-        "friendsofsymfony/oauth2-php": "~1.1",
-        "symfony/framework-bundle": "~2.8|~3.0|^4.0",
-        "symfony/security-bundle": "~2.8|~3.0|^4.0",
-        "symfony/dependency-injection": "^2.8|~3.0|^4.0",
-        "paragonie/random_compat": "^1|^2"
+  "name": "cutwise/oauth-server-bundle",
+  "type": "symfony-bundle",
+  "description": "Symfony2 OAuth Server Bundle",
+  "keywords": [
+    "oauth",
+    "oauth2",
+    "server"
+  ],
+  "homepage": "http://friendsofsymfony.github.com",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Arnaud Le Blanc",
+      "email": "arnaud.lb@gmail.com"
     },
-    "require-dev": {
-        "symfony/class-loader": "~2.8|~3.0|^4.0",
-        "symfony/yaml": "~2.8|~3.0|^4.0",
-        "symfony/form": "~2.8|~3.0|^4.0",
-        "symfony/console": "~2.8|~3.0|^4.0",
-        "symfony/templating": "~2.8|~3.0|^4.0",
-        "symfony/twig-bundle": "~2.8|~3.0|^4.0",
-        "willdurand/propel-typehintable-behavior": "^1.0.4",
-        "propel/propel1": "^1.6.5",
-        "phing/phing": "~2.4",
-        "doctrine/mongodb-odm": "~1.0",
-        "doctrine/doctrine-bundle": "~1.0",
-        "doctrine/orm": "~2.2",
-        "phpunit/phpunit": "~4.8|~5.0",
-        "symfony/phpunit-bridge": "~2.8|~3.0|^4.0",
-        "php-mock/php-mock-phpunit": "^1.1"
-    },
-    "suggest": {
-        "doctrine/doctrine-bundle": "*",
-        "doctrine/mongodb-odm-bundle": "*",
-        "propel/propel-bundle": "If you want to use Propel with Symfony2, then you will have to install the PropelBundle",
-        "willdurand/propel-typehintable-behavior": "The Typehintable behavior is useful to add type hints on generated methods, to be compliant with interfaces",
-        "symfony/form" : "Needed to be able to use the AuthorizeFormType",
-        "symfony/console": "Needed to be able to use commands"
-    },
-    "conflict": {
-        "doctrine/mongodb-odm": ">=2"
-    },
-    "autoload": {
-        "psr-4": { "FOS\\OAuthServerBundle\\": "" },
-        "exclude-from-classmap": ["/Tests/"]
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.5-dev"
-        }
+    {
+      "name": "FriendsOfSymfony Community",
+      "homepage": "https://github.com/FriendsOfSymfony/FOSOAuthServerBundle/contributors"
     }
+  ],
+  "require": {
+    "php": "^7.4|^8.0",
+    "friendsofsymfony/oauth2-php": "~1.3",
+    "symfony/framework-bundle": "^5.1",
+    "symfony/security-bundle": "^5.1",
+    "symfony/dependency-injection": "^5.1",
+    "ext-json": "*"
+  },
+  "require-dev": {
+    "symfony/yaml": "^5.1",
+    "symfony/form": "^5.1",
+    "symfony/console": "^5.1",
+    "symfony/twig-bundle": "^5.1",
+    "doctrine/doctrine-bundle": "~2.0",
+    "doctrine/orm": "~2.8",
+    "phpunit/phpunit": "~9.0",
+    "symfony/phpunit-bridge": "^5.1",
+    "php-mock/php-mock-phpunit": "^2.6"
+  },
+  "suggest": {
+    "doctrine/doctrine-bundle": "*",
+    "symfony/form": "Needed to be able to use the AuthorizeFormType",
+    "symfony/console": "Needed to be able to use commands"
+  },
+  "autoload": {
+    "psr-4": {
+      "FOS\\OAuthServerBundle\\": ""
+    },
+    "exclude-from-classmap": [
+      "/Tests/"
+    ]
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-master": "1.5-dev"
+    }
+  }
 }
