@@ -11,23 +11,23 @@
 
 namespace FOS\OAuthServerBundle\Command;
 
+use FOS\OAuthServerBundle\Model\AuthCodeManagerInterface;
+use FOS\OAuthServerBundle\Model\TokenManagerInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use FOS\OAuthServerBundle\Model\TokenManagerInterface;
-use FOS\OAuthServerBundle\Model\AuthCodeManagerInterface;
 
 class CleanCommand extends Command
 {
-    private $accessTokenManager;
-    private $refreshTokenManager;
-    private $authCodeManager;
+    private TokenManagerInterface $accessTokenManager;
+    private TokenManagerInterface $refreshTokenManager;
+    private AuthCodeManagerInterface $authCodeManager;
 
     public function __construct(
         TokenManagerInterface $accessTokenManager,
         TokenManagerInterface $refreshTokenManager,
-        AuthCodeManagerInterface $authCodeManager)
-    {
+        AuthCodeManagerInterface $authCodeManager
+    ) {
         parent::__construct();
 
         $this->accessTokenManager = $accessTokenManager;
@@ -35,32 +35,29 @@ class CleanCommand extends Command
         $this->authCodeManager = $authCodeManager;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function configure()
+    protected function configure(): void
     {
         parent::configure();
 
         $this
             ->setName('fos:oauth-server:clean')
             ->setDescription('Clean expired tokens')
-            ->setHelp(<<<EOT
+            ->setHelp(
+                <<<EOT
 The <info>%command.name%</info> command will remove expired OAuth2 tokens.
 
   <info>php %command.full_name%</info>
 EOT
-        );
+            );
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         foreach ([$this->accessTokenManager, $this->refreshTokenManager, $this->authCodeManager] as $service) {
             $result = $service->deleteExpired();
             $output->writeln(sprintf('Removed <info>%d</info> items from <comment>%s</comment> storage.', $result, get_class($service)));
         }
+
+        return 0;
     }
 }
