@@ -6,16 +6,17 @@ use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\EntityRepository;
 use FOS\OAuthServerBundle\Entity\ClientManager;
 use FOS\OAuthServerBundle\Model\ClientInterface;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Class ClientManagerTest
  * @package FOS\OAuthServerBundle\Tests\Entity
  * @author Nikola Petkanski <nikola@petkanski.com>
  */
-class ClientManagerTest extends \PHPUnit_Framework_TestCase
+class ClientManagerTest extends TestCase
 {
     /**
-     * @var \PHPUnit_Framework_MockObject_MockObject|EntityManagerInterface
+     * @var \\PHPUnit\Framework\MockObject\MockObject|EntityManagerInterface
      */
     protected $entityManager;
 
@@ -25,7 +26,7 @@ class ClientManagerTest extends \PHPUnit_Framework_TestCase
     protected $className;
 
     /**
-     * @var \PHPUnit_Framework_MockObject_MockObject|EntityRepository
+     * @var \\PHPUnit\Framework\MockObject\MockObject|EntityRepository
      */
     protected $repository;
 
@@ -34,43 +35,33 @@ class ClientManagerTest extends \PHPUnit_Framework_TestCase
      */
     protected $instance;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->entityManager = $this->getMockBuilder(EntityManagerInterface::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
         $this->repository = $this->getMockBuilder(EntityRepository::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
-        $this->className = 'RandomClassName'. \random_bytes(5);
+            ->getMock();
+        $this->className = 'RandomClassName'.\random_bytes(5);
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('getRepository')
             ->with($this->className)
-            ->willReturn($this->repository)
-        ;
+            ->willReturn($this->repository);
 
         $this->instance = new ClientManager($this->entityManager, $this->className);
 
         parent::setUp();
     }
 
-    public function testConstructWillSetParameters()
+    public function testGetClass(): void
     {
-        $this->assertAttributeSame($this->entityManager, 'em', $this->instance);
-        $this->assertAttributeSame($this->repository, 'repository', $this->instance);
-        $this->assertAttributeSame($this->className, 'class', $this->instance);
+        self::assertSame($this->className, $this->instance->getClass());
     }
 
-    public function testGetClass()
-    {
-        $this->assertSame($this->className, $this->instance->getClass());
-    }
-
-    public function testFindClientBy()
+    public function testFindClientBy(): void
     {
         $criteria = [
             \random_bytes(5),
@@ -78,60 +69,53 @@ class ClientManagerTest extends \PHPUnit_Framework_TestCase
         $randomResult = \random_bytes(5);
 
         $this->repository
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('findOneBy')
             ->with($criteria)
-            ->willReturn($randomResult)
-        ;
+            ->willReturn($randomResult);
 
-        $this->assertSame($randomResult, $this->instance->findClientBy($criteria));
+        self::assertSame($randomResult, $this->instance->findClientBy($criteria));
     }
 
-    public function testUpdateClient()
+    public function testUpdateClient(): void
     {
         $client = $this->getMockBuilder(ClientInterface::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('persist')
             ->with($client)
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('flush')
             ->with()
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
-        $this->assertNull($this->instance->updateClient($client));
+        $this->instance->updateClient($client);
     }
 
-    public function testDeleteClient()
+    public function testDeleteClient(): void
     {
         $client = $this->getMockBuilder(ClientInterface::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('remove')
             ->with($client)
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('flush')
             ->with()
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
-        $this->assertNull($this->instance->deleteClient($client));
+        $this->instance->deleteClient($client);
     }
 }
