@@ -6,17 +6,13 @@ use FOS\OAuthServerBundle\DependencyInjection\Compiler;
 use FOS\OAuthServerBundle\DependencyInjection\FOSOAuthServerExtension;
 use FOS\OAuthServerBundle\DependencyInjection\Security\Factory\OAuthFactory;
 use FOS\OAuthServerBundle\FOSOAuthServerBundle;
+use PHPUnit\Framework\TestCase;
 use Symfony\Bundle\SecurityBundle\DependencyInjection\SecurityExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Kernel;
 
-class FOSOAuthServerBundleTest extends \PHPUnit_Framework_TestCase
+class FOSOAuthServerBundleTest extends TestCase
 {
-    protected function setUp()
-    {
-        parent::setUp();
-    }
-
     public function testConstruction()
     {
         $bundle = new FOSOAuthServerBundle();
@@ -43,35 +39,33 @@ class FOSOAuthServerBundleTest extends \PHPUnit_Framework_TestCase
         $propertyReflection->setAccessible(true);
         $propertyReflection->setValue($bundle, '2.1.0');
 
-        /** @var ContainerBuilder|\PHPUnit_Framework_MockObject_MockObject $containerBuilder */
+        /** @var ContainerBuilder|\\PHPUnit\Framework\MockObject\MockObject $containerBuilder */
         $containerBuilder = $this->getMockBuilder(ContainerBuilder::class)
             ->disableOriginalConstructor()
-            ->setMethods([
-                'getExtension',
-                'addCompilerPass',
-            ])
-            ->getMock()
-        ;
+            ->setMethods(
+                [
+                    'getExtension',
+                    'addCompilerPass',
+                ]
+            )
+            ->getMock();
 
-        /** @var SecurityExtension|\PHPUnit_Framework_MockObject_MockObject $securityExtension */
+        /** @var SecurityExtension|\\PHPUnit\Framework\MockObject\MockObject $securityExtension */
         $securityExtension = $this->getMockBuilder(SecurityExtension::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $containerBuilder
             ->expects($this->at(0))
             ->method('getExtension')
             ->with('security')
-            ->willReturn($securityExtension)
-        ;
+            ->willReturn($securityExtension);
 
         $securityExtension
             ->expects($this->at(0))
             ->method('addSecurityListenerFactory')
             ->with(new OAuthFactory())
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
         $containerBuilder
             ->expects($this->at(1))
@@ -85,10 +79,9 @@ class FOSOAuthServerBundleTest extends \PHPUnit_Framework_TestCase
                 $containerBuilder,
                 $containerBuilder,
                 $containerBuilder
-            )
-        ;
+            );
 
-        $this->assertNull($bundle->build($containerBuilder));
+        $bundle->build($containerBuilder);
     }
 
     public function testBuildForSymfony20()
@@ -100,14 +93,15 @@ class FOSOAuthServerBundleTest extends \PHPUnit_Framework_TestCase
         $propertyReflection->setAccessible(true);
         $propertyReflection->setValue($bundle, '2.0.0');
 
-        /** @var ContainerBuilder|\PHPUnit_Framework_MockObject_MockObject $containerBuilder */
+        /** @var ContainerBuilder|\\PHPUnit\Framework\MockObject\MockObject $containerBuilder */
         $containerBuilder = $this->getMockBuilder(ContainerBuilder::class)
             ->disableOriginalConstructor()
-            ->setMethods([
-                'addCompilerPass',
-            ])
-            ->getMock()
-        ;
+            ->setMethods(
+                [
+                    'addCompilerPass',
+                ]
+            )
+            ->getMock();
 
         $containerBuilder
             ->expects($this->at(1))
@@ -121,9 +115,8 @@ class FOSOAuthServerBundleTest extends \PHPUnit_Framework_TestCase
                 $containerBuilder,
                 $containerBuilder,
                 $containerBuilder
-            )
-        ;
+            );
 
-        $this->assertNull($bundle->build($containerBuilder));
+        $bundle->build($containerBuilder);
     }
 }
