@@ -11,6 +11,6 @@
 
 namespace FOS\OAuthServerBundle\Tests;
 
-class TestCase extends \PHPUnit_Framework_TestCase
+class TestCase extends \PHPUnit\Framework\TestCase
 {
 }
