@@ -13,9 +13,8 @@ namespace FOS\OAuthServerBundle\DependencyInjection\Security\Factory;
 
 use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\SecurityFactoryInterface;
 use Symfony\Component\Config\Definition\Builder\NodeDefinition;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\ChildDefinition;
-use Symfony\Component\DependencyInjection\DefinitionDecorator;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
 
 /**
@@ -31,23 +30,13 @@ class OAuthFactory implements SecurityFactoryInterface
     public function create(ContainerBuilder $container, $id, $config, $userProvider, $defaultEntryPoint)
     {
         $providerId = 'security.authentication.provider.fos_oauth_server.'.$id;
-        if (class_exists(ChildDefinition::class)) {
-            $definition = new ChildDefinition('fos_oauth_server.security.authentication.provider');
-        } else {
-            $definition = new DefinitionDecorator('fos_oauth_server.security.authentication.provider');
-        }
+        $definition = new ChildDefinition('fos_oauth_server.security.authentication.provider');
         $container
             ->setDefinition($providerId, $definition)
-            ->replaceArgument(0, new Reference($userProvider))
-        ;
+            ->replaceArgument(0, new Reference($userProvider));
 
         $listenerId = 'security.authentication.listener.fos_oauth_server.'.$id;
-
-        if (class_exists(ChildDefinition::class)) {
-            $definition = new ChildDefinition('fos_oauth_server.security.authentication.listener');
-        } else {
-            $definition = new DefinitionDecorator('fos_oauth_server.security.authentication.listener');
-        }
+        $definition = new ChildDefinition('fos_oauth_server.security.authentication.listener');
         $container->setDefinition($listenerId, $definition);
 
         return array($providerId, $listenerId, 'fos_oauth_server.security.entry_point');
