@@ -12,16 +12,16 @@
 namespace FOS\OAuthServerBundle\Security\Authentication\Provider;
 
 use FOS\OAuthServerBundle\Security\Authentication\Token\OAuthToken;
+use OAuth2\OAuth2;
+use OAuth2\OAuth2AuthenticateException;
+use OAuth2\OAuth2ServerException;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\Authentication\Provider\AuthenticationProviderInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
-use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Core\Exception\AccountStatusException;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
+use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Core\User\UserCheckerInterface;
-use OAuth2\OAuth2;
-use OAuth2\OAuth2ServerException;
-use OAuth2\OAuth2AuthenticateException;
+use Symfony\Component\Security\Core\User\UserProviderInterface;
 
 /**
  * OAuthProvider class.
@@ -44,9 +44,9 @@ class OAuthProvider implements AuthenticationProviderInterface
     protected $userChecker;
 
     /**
-     * @param UserProviderInterface $userProvider  The user provider.
-     * @param OAuth2                $serverService The OAuth2 server service.
-     * @param UserCheckerInterface  $userChecker   The Symfony User Checker for Pre and Post auth checks
+     * @param UserProviderInterface $userProvider The user provider.
+     * @param OAuth2 $serverService The OAuth2 server service.
+     * @param UserCheckerInterface $userChecker The Symfony User Checker for Pre and Post auth checks
      */
     public function __construct(UserProviderInterface $userProvider, OAuth2 $serverService, UserCheckerInterface $userChecker)
     {
@@ -75,7 +75,8 @@ class OAuthProvider implements AuthenticationProviderInterface
                     try {
                         $this->userChecker->checkPreAuth($user);
                     } catch (AccountStatusException $e) {
-                        throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED,
+                        throw new OAuth2AuthenticateException(
+                            Response::HTTP_UNAUTHORIZED,
                             OAuth2::TOKEN_TYPE_BEARER,
                             $this->serverService->getVariable(OAuth2::CONFIG_WWW_REALM),
                             'access_denied',
@@ -104,7 +105,8 @@ class OAuthProvider implements AuthenticationProviderInterface
                     try {
                         $this->userChecker->checkPostAuth($user);
                     } catch (AccountStatusException $e) {
-                        throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED,
+                        throw new OAuth2AuthenticateException(
+                            Response::HTTP_UNAUTHORIZED,
                             OAuth2::TOKEN_TYPE_BEARER,
                             $this->serverService->getVariable(OAuth2::CONFIG_WWW_REALM),
                             'access_denied',
@@ -118,11 +120,6 @@ class OAuthProvider implements AuthenticationProviderInterface
                 return $token;
             }
         } catch (OAuth2ServerException $e) {
-            if (!method_exists('Symfony\Component\Security\Core\Exception\AuthenticationException', 'setToken')) {
-                // Symfony 2.1
-                throw new AuthenticationException('OAuth2 authentication failed', null, 0, $e);
-            }
-
             throw new AuthenticationException('OAuth2 authentication failed', 0, $e);
         }
 
