@@ -3,11 +3,11 @@
 namespace FOS\OAuthServerBundle\Tests\DependencyInjection\Security\Factory;
 
 use FOS\OAuthServerBundle\DependencyInjection\Security\Factory\OAuthFactory;
+use PHPUnit\Framework\TestCase;
 use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\DependencyInjection\DefinitionDecorator;
 use Symfony\Component\DependencyInjection\Reference;
 
 /**
@@ -15,53 +15,41 @@ use Symfony\Component\DependencyInjection\Reference;
  * @package FOS\OAuthServerBundle\Tests\DependencyInjection\Security\Factory
  * @author Nikola Petkanski <nikola@petkanski.com>
  */
-class OAuthFactoryTest extends \PHPUnit_Framework_TestCase
+class OAuthFactoryTest extends TestCase
 {
     /**
      * @var OAuthFactory
      */
     protected $instance;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->instance = new OAuthFactory();
 
         parent::setUp();
     }
 
-    public function testGetPosition()
+    public function testGetPosition(): void
     {
-        $this->assertSame('pre_auth', $this->instance->getPosition());
+        self::assertSame('pre_auth', $this->instance->getPosition());
     }
 
-    public function testGetKey()
+    public function testGetKey(): void
     {
-        $this->assertSame('fos_oauth', $this->instance->getKey());
+        self::assertSame('fos_oauth', $this->instance->getKey());
     }
 
-    public function testCreate()
-    {
-        if (class_exists('Symfony\Component\DependencyInjection\ChildDefinition')) {
-            return $this->useChildDefinition();
-        }
-
-        if (class_exists('Symfony\Component\DependencyInjection\DefinitionDecorator')) {
-            return $this->useDefinitionDecorator();
-        }
-
-        throw new \Exception('Neither DefinitionDecorator nor ChildDefinition exist');
-    }
-
-    protected function useDefinitionDecorator()
+    public function testCreate(): void
     {
         $container = $this->getMockBuilder(ContainerBuilder::class)
             ->disableOriginalConstructor()
-            ->setMethods([
-                'setDefinition',
+            ->setMethods(
+                [
+                    'setDefinition',
 
-            ])
-            ->getMock()
-        ;
+                ]
+            )
+            ->getMock();
         $id = '12';
         $config = [];
         $userProvider = 'mock.user.provider.service';
@@ -69,101 +57,47 @@ class OAuthFactoryTest extends \PHPUnit_Framework_TestCase
 
         $definition = $this->getMockBuilder(Definition::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $container
-            ->expects($this->exactly(2))
+            ->expects(self::exactly(2))
             ->method('setDefinition')
             ->withConsecutive(
                 [
-                    'security.authentication.provider.fos_oauth_server.'. $id,
-                    new DefinitionDecorator('fos_oauth_server.security.authentication.provider')
+                    'security.authentication.provider.fos_oauth_server.'.$id,
+                    new ChildDefinition('fos_oauth_server.security.authentication.provider'),
                 ],
                 [
-                    'security.authentication.listener.fos_oauth_server.'. $id,
-                    new DefinitionDecorator('fos_oauth_server.security.authentication.listener')
+                    'security.authentication.listener.fos_oauth_server.'.$id,
+                    new ChildDefinition('fos_oauth_server.security.authentication.listener'),
                 ]
             )
             ->willReturnOnConsecutiveCalls(
                 $definition,
                 null
-            )
-        ;
+            );
 
         $definition
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('replaceArgument')
             ->with(0, new Reference($userProvider))
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
-        $this->assertSame([
-            'security.authentication.provider.fos_oauth_server.'. $id,
-            'security.authentication.listener.fos_oauth_server.'. $id,
-            'fos_oauth_server.security.entry_point'
-        ], $this->instance->create($container, $id, $config, $userProvider, $defaultEntryPoint));
+        self::assertSame(
+            [
+                'security.authentication.provider.fos_oauth_server.'.$id,
+                'security.authentication.listener.fos_oauth_server.'.$id,
+                'fos_oauth_server.security.entry_point',
+            ],
+            $this->instance->create($container, $id, $config, $userProvider, $defaultEntryPoint)
+        );
     }
 
-    protected function useChildDefinition()
-    {
-        $container = $this->getMockBuilder(ContainerBuilder::class)
-            ->disableOriginalConstructor()
-            ->setMethods([
-                'setDefinition',
-
-            ])
-            ->getMock()
-        ;
-        $id = '12';
-        $config = [];
-        $userProvider = 'mock.user.provider.service';
-        $defaultEntryPoint = '';
-
-        $definition = $this->getMockBuilder(Definition::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
-
-        $container
-            ->expects($this->exactly(2))
-            ->method('setDefinition')
-            ->withConsecutive(
-                [
-                    'security.authentication.provider.fos_oauth_server.'. $id,
-                    new ChildDefinition('fos_oauth_server.security.authentication.provider')
-                ],
-                [
-                    'security.authentication.listener.fos_oauth_server.'. $id,
-                    new ChildDefinition('fos_oauth_server.security.authentication.listener')
-                ]
-            )
-            ->willReturnOnConsecutiveCalls(
-                $definition,
-                null
-            )
-        ;
-
-        $definition
-            ->expects($this->once())
-            ->method('replaceArgument')
-            ->with(0, new Reference($userProvider))
-            ->willReturn(null)
-        ;
-
-        $this->assertSame([
-            'security.authentication.provider.fos_oauth_server.'. $id,
-            'security.authentication.listener.fos_oauth_server.'. $id,
-            'fos_oauth_server.security.entry_point'
-        ], $this->instance->create($container, $id, $config, $userProvider, $defaultEntryPoint));
-    }
-
-    public function testAddConfigurationDoesNothing()
+    public function testAddConfigurationDoesNothing(): void
     {
         $nodeDefinition = $this->getMockBuilder(NodeDefinition::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
-        $this->assertNull($this->instance->addConfiguration($nodeDefinition));
+            ->getMock();
+        $this->instance->addConfiguration($nodeDefinition);
     }
 }
