@@ -2,13 +2,15 @@
 
 namespace FOS\OAuthServerBundle\Tests\Entity;
 
-use Doctrine\Common\Persistence\ObjectRepository;
 use Doctrine\ORM\AbstractQuery;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\QueryBuilder;
+use Doctrine\Persistence\ObjectRepository;
 use FOS\OAuthServerBundle\Entity\AuthCodeManager;
 use FOS\OAuthServerBundle\Model\AuthCodeInterface;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @group time-sensitive
@@ -17,10 +19,10 @@ use FOS\OAuthServerBundle\Model\AuthCodeInterface;
  * @package FOS\OAuthServerBundle\Tests\Entity
  * @author Nikola Petkanski <nikola@petkanski.com>
  */
-class AuthCodeManagerTest extends \PHPUnit_Framework_TestCase
+class AuthCodeManagerTest extends TestCase
 {
     /**
-     * @var \PHPUnit_Framework_MockObject_MockObject|EntityManagerInterface
+     * @var MockObject|EntityManagerInterface
      */
     protected $entityManager;
 
@@ -34,175 +36,147 @@ class AuthCodeManagerTest extends \PHPUnit_Framework_TestCase
      */
     protected $instance;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->entityManager = $this->getMockBuilder(EntityManagerInterface::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
-        $this->className = 'TestClassName' . \random_bytes(5);
+            ->getMock();
+        $this->className = 'TestClassName'.random_bytes(5);
 
         $this->instance = new AuthCodeManager($this->entityManager, $this->className);
 
         parent::setUp();
     }
 
-    public function testConstructWillSetParameters()
+    public function testGetClassWillReturnClassName(): void
     {
-        $this->assertAttributeSame($this->entityManager, 'em', $this->instance);
-        $this->assertAttributeSame($this->className, 'class', $this->instance);
+        self::assertSame($this->className, $this->instance->getClass());
     }
 
-    public function testGetClassWillReturnClassName()
-    {
-        $this->assertSame($this->className, $this->instance->getClass());
-    }
-
-    public function testFindAuthCodeBy()
+    public function testFindAuthCodeBy(): void
     {
         $repository = $this->getMockBuilder(ObjectRepository::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('getRepository')
             ->with($this->className)
-            ->willReturn($repository)
-        ;
+            ->willReturn($repository);
 
-        $criteria = [
-            \random_bytes(10)
-        ];
-        $randomResult = \random_bytes(10);
+        $criteria = [random_bytes(10)];
+        $randomResult = random_bytes(10);
 
         $repository
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('findOneBy')
             ->with($criteria)
-            ->willReturn($randomResult)
-        ;
+            ->willReturn($randomResult);
 
-        $this->assertSame($randomResult, $this->instance->findAuthCodeBy($criteria));
+        self::assertSame($randomResult, $this->instance->findAuthCodeBy($criteria));
     }
 
-    public function testUpdateAuthCode()
+    public function testUpdateAuthCode(): void
     {
         $authCode = $this->getMockBuilder(AuthCodeInterface::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('persist')
             ->with($authCode)
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('flush')
             ->with()
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
-        $this->assertNull($this->instance->updateAuthCode($authCode));
+        $this->instance->updateAuthCode($authCode);
     }
 
-    public function testDeleteAuthCode()
+    public function testDeleteAuthCode(): void
     {
         $authCode = $this->getMockBuilder(AuthCodeInterface::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('remove')
             ->with($authCode)
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('flush')
             ->with()
-            ->willReturn(null)
-        ;
+            ->willReturn(null);
 
-        $this->assertNull($this->instance->deleteAuthCode($authCode));
+        $this->instance->deleteAuthCode($authCode);
     }
 
-    public function testDeleteExpired()
+    public function testDeleteExpired(): void
     {
-        $randomResult = \random_bytes(10);
+        $randomResult = random_bytes(10);
 
         $queryBuilder = $this->getMockBuilder(QueryBuilder::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $repository = $this->getMockBuilder(EntityRepository::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $this->entityManager
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('getRepository')
             ->with($this->className)
-            ->willReturn($repository)
-        ;
+            ->willReturn($repository);
 
         $repository
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('createQueryBuilder')
             ->with('a')
-            ->willReturn($queryBuilder)
-        ;
+            ->willReturn($queryBuilder);
 
         $queryBuilder
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('delete')
             ->with()
-            ->willReturn($queryBuilder)
-        ;
+            ->willReturn($queryBuilder);
 
         $queryBuilder
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('where')
             ->with('a.expiresAt < ?1')
-            ->willReturn($queryBuilder)
-        ;
+            ->willReturn($queryBuilder);
 
         $queryBuilder
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('setParameters')
             ->with([1 => time()])
-            ->willReturn($queryBuilder)
-        ;
+            ->willReturn($queryBuilder);
 
         $query = $this->getMockBuilder(AbstractQuery::class)
             ->disableOriginalConstructor()
-            ->getMock()
-        ;
+            ->getMock();
 
         $queryBuilder
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('getQuery')
             ->with()
-            ->willReturn($query)
-        ;
+            ->willReturn($query);
 
         $query
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('execute')
             ->with()
-            ->willReturn($randomResult)
-        ;
+            ->willReturn($randomResult);
 
-        $this->assertSame($randomResult, $this->instance->deleteExpired());
+        self::assertSame($randomResult, $this->instance->deleteExpired());
     }
 }
