@@ -12,13 +12,24 @@
 namespace FOS\OAuth2ServiceBundle\Tests\Storage;
 
 use FOS\OAuthServerBundle\Model\AccessToken;
-use FOS\OAuthServerBundle\Model\RefreshToken;
+use FOS\OAuthServerBundle\Model\AccessTokenManagerInterface;
 use FOS\OAuthServerBundle\Model\AuthCode;
+use FOS\OAuthServerBundle\Model\AuthCodeInterface;
+use FOS\OAuthServerBundle\Model\AuthCodeManagerInterface;
 use FOS\OAuthServerBundle\Model\Client;
+use FOS\OAuthServerBundle\Model\ClientManagerInterface;
+use FOS\OAuthServerBundle\Model\RefreshToken;
+use FOS\OAuthServerBundle\Model\RefreshTokenManagerInterface;
+use FOS\OAuthServerBundle\Storage\GrantExtensionInterface;
 use FOS\OAuthServerBundle\Storage\OAuthStorage;
+use OAuth2\Model\IOAuth2Client;
+use PHPUnit\Framework\TestCase;
+use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
+use Symfony\Component\Security\Core\Encoder\PasswordEncoderInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\Security\Core\User\UserProviderInterface;
 
-class OAuthStorageTest extends \PHPUnit_Framework_TestCase
+class OAuthStorageTest extends TestCase
 {
     protected $clientManager;
 
@@ -34,463 +45,478 @@ class OAuthStorageTest extends \PHPUnit_Framework_TestCase
 
     protected $storage;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->clientManager = $this->getMockBuilder('FOS\OAuthServerBundle\Model\ClientManagerInterface')
+        $this->clientManager = $this->getMockBuilder(ClientManagerInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $this->accessTokenManager = $this->getMockBuilder('FOS\OAuthServerBundle\Model\AccessTokenManagerInterface')
+        $this->accessTokenManager = $this->getMockBuilder(AccessTokenManagerInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $this->refreshTokenManager = $this->getMockBuilder('FOS\OAuthServerBundle\Model\RefreshTokenManagerInterface')
+        $this->refreshTokenManager = $this->getMockBuilder(RefreshTokenManagerInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $this->authCodeManager = $this->getMockBuilder('FOS\OAuthServerBundle\Model\AuthCodeManagerInterface')
+        $this->authCodeManager = $this->getMockBuilder(AuthCodeManagerInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $this->userProvider = $this->getMockBuilder('Symfony\Component\Security\Core\User\UserProviderInterface')
+        $this->userProvider = $this->getMockBuilder(UserProviderInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $this->encoderFactory = $this->getMockBuilder('Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface')
+        $this->encoderFactory = $this->getMockBuilder(EncoderFactoryInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $this->storage = new OAuthStorage($this->clientManager, $this->accessTokenManager, $this->refreshTokenManager, $this->authCodeManager, $this->userProvider, $this->encoderFactory);
+        $this->storage = new OAuthStorage(
+            $this->clientManager,
+            $this->accessTokenManager,
+            $this->refreshTokenManager,
+            $this->authCodeManager,
+            $this->userProvider,
+            $this->encoderFactory
+        );
     }
 
-    public function testGetClientReturnsClientWithGivenId()
+    public function testGetClientReturnsClientWithGivenId(): void
     {
         $client = new Client();
 
-        $this->clientManager->expects($this->once())
+        $this->clientManager->expects(self::once())
             ->method('findClientByPublicId')
             ->with('123_abc')
-            ->will($this->returnValue($client));
+            ->willReturn($client);
 
-        $this->assertSame($client, $this->storage->getClient('123_abc'));
+        self::assertSame($client, $this->storage->getClient('123_abc'));
     }
 
-    public function testGetClientReturnsNullIfNotExists()
+    public function testGetClientReturnsNullIfNotExists(): void
     {
-        $client = new Client();
-
-        $this->clientManager->expects($this->once())
+        $this->clientManager->expects(self::once())
             ->method('findClientByPublicId')
             ->with('123_abc')
-            ->will($this->returnValue(null));
+            ->willReturn(null);
 
-        $this->assertNull($this->storage->getClient('123_abc'));
+        self::assertNull($this->storage->getClient('123_abc'));
     }
 
-    public function testCheckClientCredentialsThrowsIfInvalidClientClass()
+    public function testCheckClientCredentialsThrowsIfInvalidClientClass(): void
     {
-        $client = $this->getMockBuilder('OAuth2\Model\IOAuth2Client')
+        $client = $this->getMockBuilder(IOAuth2Client::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
         $this->storage->checkClientCredentials($client, 'dummy');
     }
 
-    public function testCheckClientCredentialsReturnsTrueOnValidCredentials()
+    public function testCheckClientCredentialsReturnsTrueOnValidCredentials(): void
     {
         $client = new Client();
         $client->setSecret('dummy');
 
-        $this->assertTrue($this->storage->checkClientCredentials($client, 'dummy'));
+        self::assertTrue($this->storage->checkClientCredentials($client, 'dummy'));
     }
 
-    public function testCheckClientCredentialsReturnsFalseOnValidCredentials()
+    public function testCheckClientCredentialsReturnsFalseOnValidCredentials(): void
     {
         $client = new Client();
         $client->setSecret('dummy');
 
-        $this->assertFalse($this->storage->checkClientCredentials($client, 'passe'));
+        self::assertFalse($this->storage->checkClientCredentials($client, 'passe'));
     }
 
-    public function testGetAccessTokenReturnsAccessTokenWithGivenId()
+    public function testGetAccessTokenReturnsAccessTokenWithGivenId(): void
     {
         $token = new AccessToken();
 
-        $this->accessTokenManager->expects($this->once())
+        $this->accessTokenManager->expects(self::once())
             ->method('findTokenByToken')
             ->with('123_abc')
-            ->will($this->returnValue($token));
+            ->willReturn($token);
 
-        $this->assertSame($token, $this->storage->getAccessToken('123_abc'));
+        self::assertSame($token, $this->storage->getAccessToken('123_abc'));
     }
 
-    public function testGetAccessTokenReturnsNullIfNotExists()
+    public function testGetAccessTokenReturnsNullIfNotExists(): void
     {
-        $token = new AccessToken();
-
-        $this->accessTokenManager->expects($this->once())
+        $this->accessTokenManager->expects(self::once())
             ->method('findTokenByToken')
             ->with('123_abc')
-            ->will($this->returnValue(null));
+            ->willReturn(null);
 
-        $this->assertNull($this->storage->getAccessToken('123_abc'));
+        self::assertNull($this->storage->getAccessToken('123_abc'));
     }
 
-    public function testCreateAccessTokenThrowsOnInvalidClientClass()
+    public function testCreateAccessTokenThrowsOnInvalidClientClass(): void
     {
-        $client = $this->getMockBuilder('OAuth2\Model\IOAuth2Client')
+        $client = $this->getMockBuilder(IOAuth2Client::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
         $this->storage->createAccessToken('foo', $client, new User(42), 1, 'foo bar');
     }
 
-    public function testCreateAccessToken()
+    public function testCreateAccessToken(): void
     {
         $savedToken = null;
 
-        $this->accessTokenManager->expects($this->once())
+        $this->accessTokenManager->expects(self::once())
             ->method('createToken')
             ->with()
-            ->will($this->returnValue(new AccessToken()));
-        $this->accessTokenManager->expects($this->once())
+            ->willReturn(new AccessToken());
+        $this->accessTokenManager->expects(self::once())
             ->method('updateToken')
-            ->will($this->returnCallback(function ($token) use (&$savedToken) {
-                $savedToken = $token;
-            }));
+            ->willReturnCallback(
+                function ($token) use (&$savedToken) {
+                    $savedToken = $token;
+                }
+            );
 
         $client = new Client();
         $user = new User(42);
 
         $token = $this->storage->createAccessToken('foo', $client, $user, 1, 'foo bar');
 
-        $this->assertEquals($token, $savedToken);
+        self::assertEquals($token, $savedToken);
 
-        $this->assertSame('foo', $token->getToken());
-        $this->assertSame($client, $token->getClient());
-        $this->assertSame($user, $token->getData());
-        $this->assertSame($user, $token->getUser());
-        $this->assertSame(1, $token->getExpiresAt());
-        $this->assertSame('foo bar', $token->getScope());
+        self::assertSame('foo', $token->getToken());
+        self::assertSame($client, $token->getClient());
+        self::assertSame($user, $token->getData());
+        self::assertSame($user, $token->getUser());
+        self::assertSame(1, $token->getExpiresAt());
+        self::assertSame('foo bar', $token->getScope());
     }
 
-    public function testCreateAccessTokenWithoutUser()
+    public function testCreateAccessTokenWithoutUser(): void
     {
         $savedToken = null;
 
-        $this->accessTokenManager->expects($this->once())
+        $this->accessTokenManager->expects(self::once())
             ->method('createToken')
             ->with()
-            ->will($this->returnValue(new AccessToken()));
-        $this->accessTokenManager->expects($this->once())
+            ->willReturn(new AccessToken());
+        $this->accessTokenManager->expects(self::once())
             ->method('updateToken')
-            ->will($this->returnCallback(function ($token) use (&$savedToken) {
-            $savedToken = $token;
-        }));
+            ->willReturnCallback(
+                function ($token) use (&$savedToken) {
+                    $savedToken = $token;
+                }
+            );
 
         $client = new Client();
         $user = null;
 
         $token = $this->storage->createAccessToken('foo', $client, $user, 1, 'foo bar');
 
-        $this->assertEquals($token, $savedToken);
+        self::assertEquals($token, $savedToken);
     }
 
-    public function testGetRefreshTokenReturnsRefreshTokenWithGivenId()
+    public function testGetRefreshTokenReturnsRefreshTokenWithGivenId(): void
     {
         $token = new RefreshToken();
 
-        $this->refreshTokenManager->expects($this->once())
+        $this->refreshTokenManager->expects(self::once())
             ->method('findTokenByToken')
             ->with('123_abc')
-            ->will($this->returnValue($token));
+            ->willReturn($token);
 
-        $this->assertSame($token, $this->storage->getRefreshToken('123_abc'));
+        self::assertSame($token, $this->storage->getRefreshToken('123_abc'));
     }
 
-    public function testGetRefreshTokenReturnsNullIfNotExists()
+    public function testGetRefreshTokenReturnsNullIfNotExists(): void
     {
-        $this->refreshTokenManager->expects($this->once())
+        $this->refreshTokenManager->expects(self::once())
             ->method('findTokenByToken')
             ->with('123_abc')
-            ->will($this->returnValue(null));
+            ->willReturn(null);
 
-        $this->assertNull($this->storage->getRefreshToken('123_abc'));
+        self::assertNull($this->storage->getRefreshToken('123_abc'));
     }
 
-    public function testCreateRefreshTokenThrowsOnInvalidClientClass()
+    public function testCreateRefreshTokenThrowsOnInvalidClientClass(): void
     {
-        $client = $this->getMockBuilder('OAuth2\Model\IOAuth2Client')
+        $client = $this->getMockBuilder(IOAuth2Client::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
         $this->storage->createRefreshToken('foo', $client, 42, 1, 'foo bar');
     }
 
-    public function testCreateRefreshToken()
+    public function testCreateRefreshToken(): void
     {
         $savedToken = null;
 
-        $this->refreshTokenManager->expects($this->once())
+        $this->refreshTokenManager->expects(self::once())
             ->method('createToken')
             ->with()
-            ->will($this->returnValue(new RefreshToken()));
-        $this->refreshTokenManager->expects($this->once())
+            ->willReturn(new RefreshToken());
+        $this->refreshTokenManager->expects(self::once())
             ->method('updateToken')
-            ->will($this->returnCallback(function ($token) use (&$savedToken) {
-                $savedToken = $token;
-            }));
+            ->willReturnCallback(
+                function ($token) use (&$savedToken) {
+                    $savedToken = $token;
+                }
+            );
 
         $client = new Client();
         $user = new User(42);
 
         $token = $this->storage->createRefreshToken('foo', $client, $user, 1, 'foo bar');
 
-        $this->assertEquals($token, $savedToken);
+        self::assertEquals($token, $savedToken);
 
-        $this->assertSame('foo', $token->getToken());
-        $this->assertSame($client, $token->getClient());
-        $this->assertSame($user, $token->getData());
-        $this->assertSame($user, $token->getUser());
-        $this->assertSame(1, $token->getExpiresAt());
-        $this->assertSame('foo bar', $token->getScope());
+        self::assertSame('foo', $token->getToken());
+        self::assertSame($client, $token->getClient());
+        self::assertSame($user, $token->getData());
+        self::assertSame($user, $token->getUser());
+        self::assertSame(1, $token->getExpiresAt());
+        self::assertSame('foo bar', $token->getScope());
     }
 
-    public function testCreateRefreshTokenWithoutUser()
+    public function testCreateRefreshTokenWithoutUser(): void
     {
         $savedToken = null;
 
-        $this->refreshTokenManager->expects($this->once())
+        $this->refreshTokenManager->expects(self::once())
             ->method('createToken')
             ->with()
-            ->will($this->returnValue(new RefreshToken()));
-        $this->refreshTokenManager->expects($this->once())
+            ->willReturn(new RefreshToken());
+        $this->refreshTokenManager->expects(self::once())
             ->method('updateToken')
-            ->will($this->returnCallback(function ($token) use (&$savedToken) {
-            $savedToken = $token;
-        }));
+            ->willReturnCallback(
+                function ($token) use (&$savedToken) {
+                    $savedToken = $token;
+                }
+            );
 
         $client = new Client();
         $user = null;
 
         $token = $this->storage->createRefreshToken('foo', $client, $user, 1, 'foo bar');
 
-        $this->assertEquals($token, $savedToken);
+        self::assertEquals($token, $savedToken);
     }
 
-    public function testCheckRestrictedGrantTypeThrowsOnInvalidClientClass()
+    public function testCheckRestrictedGrantTypeThrowsOnInvalidClientClass(): void
     {
-        $client = $this->getMockBuilder('OAuth2\Model\IOAuth2Client')
+        $client = $this->getMockBuilder(IOAuth2Client::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
 
         $this->storage->checkRestrictedGrantType($client, 'foo');
     }
 
-    public function testCheckRestrictedGrantType()
+    public function testCheckRestrictedGrantType(): void
     {
         $client = new Client();
         $client->setAllowedGrantTypes(array('foo', 'bar'));
 
-        $this->assertTrue($this->storage->checkRestrictedGrantType($client, 'foo'));
-        $this->assertTrue($this->storage->checkRestrictedGrantType($client, 'bar'));
-        $this->assertFalse($this->storage->checkRestrictedGrantType($client, 'baz'));
+        self::assertTrue($this->storage->checkRestrictedGrantType($client, 'foo'));
+        self::assertTrue($this->storage->checkRestrictedGrantType($client, 'bar'));
+        self::assertFalse($this->storage->checkRestrictedGrantType($client, 'baz'));
     }
 
-    public function testCheckUserCredentialsThrowsOnInvalidClientClass()
+    public function testCheckUserCredentialsThrowsOnInvalidClientClass(): void
     {
-        $client = $this->getMockBuilder('OAuth2\Model\IOAuth2Client')
+        $client = $this->getMockBuilder(IOAuth2Client::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
 
         $this->storage->checkUserCredentials($client, 'Joe', 'baz');
     }
 
-    public function testCheckUserCredentialsCatchesAuthenticationExceptions()
+    public function testCheckUserCredentialsCatchesAuthenticationExceptions(): void
     {
         $client = new Client();
 
         $result = $this->storage->checkUserCredentials($client, 'Joe', 'baz');
 
-        $this->assertFalse($result);
+        self::assertFalse($result);
     }
 
-    public function testCheckUserCredentialsReturnsTrueOnValidCredentials()
+    public function testCheckUserCredentialsReturnsTrueOnValidCredentials(): void
     {
         $client = new Client();
-        $user = $this->getMockBuilder('Symfony\Component\Security\Core\User\UserInterface')
+        $user = $this->getMockBuilder(UserInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $user->expects($this->once())
-            ->method('getPassword')->with()->will($this->returnValue('foo'));
-        $user->expects($this->once())
-            ->method('getSalt')->with()->will($this->returnValue('bar'));
+        $user->expects(self::once())
+            ->method('getPassword')->with()->willReturn('foo');
+        $user->expects(self::once())
+            ->method('getSalt')->with()->willReturn('bar');
 
-        $encoder = $this->getMockBuilder('Symfony\Component\Security\Core\Encoder\PasswordEncoderInterface')
+        $encoder = $this->getMockBuilder(PasswordEncoderInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $encoder->expects($this->once())
+        $encoder->expects(self::once())
             ->method('isPasswordValid')
             ->with('foo', 'baz', 'bar')
-            ->will($this->returnValue(true));
+            ->willReturn(true);
 
-        $this->userProvider->expects($this->once())
+        $this->userProvider->expects(self::once())
             ->method('loadUserByUsername')
             ->with('Joe')
-            ->will($this->returnValue($user));
+            ->willReturn($user);
 
-        $this->encoderFactory->expects($this->once())
+        $this->encoderFactory->expects(self::once())
             ->method('getEncoder')
             ->with($user)
-            ->will($this->returnValue($encoder));
+            ->willReturn($encoder);
 
-        $this->assertSame(array(
-            'data' => $user,
-        ), $this->storage->checkUserCredentials($client, 'Joe', 'baz'));
+        self::assertSame(
+            array(
+                'data' => $user,
+            ),
+            $this->storage->checkUserCredentials($client, 'Joe', 'baz')
+        );
     }
 
-    public function testCheckUserCredentialsReturnsFalseOnInvalidCredentials()
+    public function testCheckUserCredentialsReturnsFalseOnInvalidCredentials(): void
     {
         $client = new Client();
-        $user = $this->getMockBuilder('Symfony\Component\Security\Core\User\UserInterface')
+        $user = $this->getMockBuilder(UserInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $user->expects($this->once())
-            ->method('getPassword')->with()->will($this->returnValue('foo'));
-        $user->expects($this->once())
-            ->method('getSalt')->with()->will($this->returnValue('bar'));
+        $user->expects(self::once())
+            ->method('getPassword')->with()->willReturn('foo');
+        $user->expects(self::once())
+            ->method('getSalt')->with()->willReturn('bar');
 
-        $encoder = $this->getMockBuilder('Symfony\Component\Security\Core\Encoder\PasswordEncoderInterface')
+        $encoder = $this->getMockBuilder(PasswordEncoderInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $encoder->expects($this->once())
+        $encoder->expects(self::once())
             ->method('isPasswordValid')
             ->with('foo', 'baz', 'bar')
-            ->will($this->returnValue(false));
+            ->willReturn(false);
 
-        $this->userProvider->expects($this->once())
+        $this->userProvider->expects(self::once())
             ->method('loadUserByUsername')
             ->with('Joe')
-            ->will($this->returnValue($user));
+            ->willReturn($user);
 
-        $this->encoderFactory->expects($this->once())
+        $this->encoderFactory->expects(self::once())
             ->method('getEncoder')
             ->with($user)
-            ->will($this->returnValue($encoder));
+            ->willReturn($encoder);
 
-        $this->assertFalse($this->storage->checkUserCredentials($client, 'Joe', 'baz'));
+        self::assertFalse($this->storage->checkUserCredentials($client, 'Joe', 'baz'));
     }
 
-    public function testCheckUserCredentialsReturnsFalseIfUserNotExist()
+    public function testCheckUserCredentialsReturnsFalseIfUserNotExist(): void
     {
         $client = new Client();
 
-        $this->userProvider->expects($this->once())
+        $this->userProvider->expects(self::once())
             ->method('loadUserByUsername')
             ->with('Joe')
-            ->will($this->returnValue(null));
+            ->willReturn(null);
 
-        $this->assertFalse($this->storage->checkUserCredentials($client, 'Joe', 'baz'));
+        self::assertFalse($this->storage->checkUserCredentials($client, 'Joe', 'baz'));
     }
 
-    public function testCreateAuthCodeThrowsOnInvalidClientClass()
+    public function testCreateAuthCodeThrowsOnInvalidClientClass(): void
     {
-        $client = $this->getMockBuilder('OAuth2\Model\IOAuth2Client')
+        $client = $this->getMockBuilder(IOAuth2Client::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
         $this->storage->createAuthCode('foo', $client, 42, 'http://www.example.com/', 1, 'foo bar');
     }
 
-    public function testCreateAuthCode()
+    public function testCreateAuthCode(): void
     {
         $savedCode = null;
 
-        $this->authCodeManager->expects($this->once())
+        $this->authCodeManager->expects(self::once())
             ->method('createAuthCode')
             ->with()
-            ->will($this->returnValue(new AuthCode()));
-        $this->authCodeManager->expects($this->once())
+            ->willReturn(new AuthCode());
+        $this->authCodeManager->expects(self::once())
             ->method('updateAuthCode')
-            ->will($this->returnCallback(function ($code) use (&$savedCode) {
-                $savedCode = $code;
-            }));
+            ->willReturnCallback(
+                function ($code) use (&$savedCode) {
+                    $savedCode = $code;
+                }
+            );
 
         $client = new Client();
         $user = new User(42);
 
         $code = $this->storage->createAuthCode('foo', $client, $user, 'http://www.example.com/', 1, 'foo bar');
 
-        $this->assertEquals($code, $savedCode);
+        self::assertEquals($code, $savedCode);
 
-        $this->assertSame('foo', $code->getToken());
-        $this->assertSame($client, $code->getClient());
-        $this->assertSame($user, $code->getData());
-        $this->assertSame($user, $code->getUser());
-        $this->assertSame(1, $code->getExpiresAt());
-        $this->assertSame('foo bar', $code->getScope());
+        self::assertSame('foo', $code->getToken());
+        self::assertSame($client, $code->getClient());
+        self::assertSame($user, $code->getData());
+        self::assertSame($user, $code->getUser());
+        self::assertSame(1, $code->getExpiresAt());
+        self::assertSame('foo bar', $code->getScope());
     }
 
-    public function testGetAuthCodeReturnsAuthCodeWithGivenId()
+    public function testGetAuthCodeReturnsAuthCodeWithGivenId(): void
     {
         $code = new AuthCode();
 
-        $this->authCodeManager->expects($this->once())
+        $this->authCodeManager->expects(self::once())
             ->method('findAuthCodeByToken')
             ->with('123_abc')
-            ->will($this->returnValue($code));
+            ->willReturn($code);
 
-        $this->assertSame($code, $this->storage->getAuthCode('123_abc'));
+        self::assertSame($code, $this->storage->getAuthCode('123_abc'));
     }
 
-    public function testGetAuthCodeReturnsNullIfNotExists()
+    public function testGetAuthCodeReturnsNullIfNotExists(): void
     {
-        $this->authCodeManager->expects($this->once())
+        $this->authCodeManager->expects(self::once())
             ->method('findAuthCodeByToken')
             ->with('123_abc')
-            ->will($this->returnValue(null));
+            ->willReturn(null);
 
-        $this->assertNull($this->storage->getAuthCode('123_abc'));
+        self::assertNull($this->storage->getAuthCode('123_abc'));
     }
 
-    public function testValidGrantExtension()
+    public function testValidGrantExtension(): void
     {
-        $grantExtension = $this->getMockBuilder('FOS\OAuthServerBundle\Storage\GrantExtensionInterface')
+        $grantExtension = $this->getMockBuilder(GrantExtensionInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
         $grantExtension
-            ->expects($this->once())
+            ->expects(self::once())
             ->method('checkGrantExtension')
-            ->will($this->returnValue(true))
-        ;
+            ->willReturn(true);
         $this->storage->setGrantExtension('https://friendsofsymfony.com/grants/foo', $grantExtension);
 
-        $client = $this->getMockBuilder('OAuth2\Model\IOAuth2Client')
+        $client = $this->getMockBuilder(IOAuth2Client::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $this->assertTrue($this->storage->checkGrantExtension($client, 'https://friendsofsymfony.com/grants/foo', array(), array()));
+        self::assertTrue($this->storage->checkGrantExtension($client, 'https://friendsofsymfony.com/grants/foo', array(), array()));
     }
 
     /**
      * @expectedException \OAuth2\OAuth2ServerException
      */
-    public function testInvalidGrantExtension()
+    public function testInvalidGrantExtension(): void
     {
-        $client = $this->getMockBuilder('OAuth2\Model\IOAuth2Client')
+        $client = $this->getMockBuilder(IOAuth2Client::class)
             ->disableOriginalConstructor()
             ->getMock();
         $this->storage->checkGrantExtension($client, 'https://friendsofsymfony.com/grants/bar', array(), array());
     }
 
-    public function testDoubleSetGrantExtension()
+    public function testDoubleSetGrantExtension(): void
     {
-        $grantExtension = $this->getMockBuilder('FOS\OAuthServerBundle\Storage\GrantExtensionInterface')
+        $grantExtension = $this->getMockBuilder(GrantExtensionInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $grantExtension2 = $this->getMockBuilder('FOS\OAuthServerBundle\Storage\GrantExtensionInterface')
+        $grantExtension2 = $this->getMockBuilder(GrantExtensionInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
         $this->storage->setGrantExtension($uri = 'https://friendsofsymfony.com/grants/foo', $grantExtension);
@@ -501,41 +527,37 @@ class OAuthStorageTest extends \PHPUnit_Framework_TestCase
         $grantExtensionsProperty->setAccessible(true);
         $grantExtensions = $grantExtensionsProperty->getValue($this->storage);
 
-        $this->assertEquals($grantExtension2, $grantExtensions[$uri]);
+        self::assertEquals($grantExtension2, $grantExtensions[$uri]);
     }
 
-    public function testMarkAuthCodeAsUsedIfAuthCodeFound()
+    public function testMarkAuthCodeAsUsedIfAuthCodeFound(): void
     {
-        $authCode = $this->getMockBuilder('FOS\OAuthServerBundle\Model\AuthCodeInterface')
+        $authCode = $this->getMockBuilder(AuthCodeInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $this->authCodeManager->expects($this->atLeastOnce())
+        $this->authCodeManager->expects(self::atLeastOnce())
             ->method('findAuthCodeByToken')
             ->with('123_abc')
-            ->will($this->returnValue($authCode))
-        ;
+            ->willReturn($authCode);
 
-        $this->authCodeManager->expects($this->atLeastOnce())
+        $this->authCodeManager->expects(self::atLeastOnce())
             ->method('deleteAuthCode')
             ->with($authCode)
-            ->will($this->returnValue(null))
-        ;
+            ->willReturn(null);
 
         $this->storage->markAuthCodeAsUsed('123_abc');
     }
 
-    public function testMarkAuthCodeAsUsedIfAuthCodeNotFound()
+    public function testMarkAuthCodeAsUsedIfAuthCodeNotFound(): void
     {
-        $this->authCodeManager->expects($this->atLeastOnce())
+        $this->authCodeManager->expects(self::atLeastOnce())
             ->method('findAuthCodeByToken')
             ->with('123_abc')
-            ->will($this->returnValue(null))
-        ;
+            ->willReturn(null);
 
         $this->authCodeManager->expects($this->never())
-            ->method('deleteAuthCode')
-        ;
+            ->method('deleteAuthCode');
 
         $this->storage->markAuthCodeAsUsed('123_abc');
     }
